Made GroupScrollArea a separate class
