+++ shortcut_composer/templates/pie_menu_utils/settings_gui/components/__init__.py
@@ -1,9 +1,11 @@
 from .group_combo_box import GroupComboBox
 from .group_fetcher import GroupFetcher, EnumGroupFetcher, PresetGroupFetcher
+from .group_scroll_area import GroupScrollArea
 
 __all__ = [
     "GroupComboBox",
     "GroupFetcher",
     "EnumGroupFetcher",
-    "PresetGroupFetcher"
+    "PresetGroupFetcher",
+    "GroupScrollArea",
 ]
