+++ shortcut_composer/templates/pie_menu_utils/settings_gui/enum_group_pie_settings.py
@@ -3,45 +3,11 @@
 
 from api_krita.enums.helpers import EnumGroup
 from core_components import Controller
-from config_system import Field
 from ..label import Label
 from ..pie_style import PieStyle
 from ..pie_config import NonPresetPieConfig
 from .pie_settings import PieSettings
-from .scroll_area import ScrollArea
-from .components import GroupComboBox, EnumGroupFetcher, GroupFetcher
-
-
-class GroupScrollArea(ScrollArea):
-    def __init__(
-        self,
-        controller: Controller[EnumGroup],
-        fetcher: GroupFetcher,
-        style: PieStyle,
-        columns: int,
-        field: Field,
-        parent=None
-    ) -> None:
-        super().__init__(style, columns, parent)
-        self._controller = controller
-        self._field = field
-        self._fetcher = fetcher
-        self.group_chooser = GroupComboBox(
-            config_field=self._field,
-            group_fetcher=self._fetcher,
-            additional_fields=["All"])
-        self.group_chooser.widget.currentTextChanged.connect(
-            self._display_group)
-        self._layout.insertWidget(0, self.group_chooser.widget)
-        self._display_group()
-
-    def _display_group(self) -> None:
-        """Update preset widgets according to tag selected in combobox."""
-        picked_group = self.group_chooser.widget.currentText()
-        values = self._fetcher.get_values(picked_group)
-        self.replace_handled_labels(self._fetcher.create_labels(values))
-        self._apply_search_bar_filter()
-        self.group_chooser.save()
+from .components import EnumGroupFetcher, GroupScrollArea
 
 
 class EnumGroupPieSettings(PieSettings):
@@ -64,7 +30,8 @@ class EnumGroupPieSettings(PieSettings):
             fetcher=EnumGroupFetcher(self._controller),
             style=self._style,
             columns=3,
-            field=self._config.field("Last tag selected", "All"))
+            field=self._config.field("Last tag selected", "All"),
+            additional_fields=["All"])
         self._action_values.replace_handled_labels(labels)
         self._tab_holder.insertTab(1, self._action_values, "Values")
         self._tab_holder.setCurrentIndex(1)
@@ -75,8 +42,3 @@ class EnumGroupPieSettings(PieSettings):
     def _refresh_draggable(self) -> None:
         """Make all values currently used in pie undraggable and disabled."""
         self._action_values.mark_used_values(self._config.values())
-
-    # def create_labels(self, values: List[Enum]) -> List[Label[Enum]]:
-    #     """Create labels from list of preset names."""
-    #     labels = [Label.from_value(v, self._controller) for v in values]
-    #     return [label for label in labels if label is not None]
